requirements.txt
--index-url https://mirrors.aliyun.com/pypi/simple/

langchain>=0.3.14
langchain-community>=0.3.14
pydantic>=2.10.5
openai>=1.3.7
zhipuai>=1.0.7
python-docx>=1.0.1
python-pptx>=0.6.21
openpyxl>=3.1.2
pytesseract>=0.3.10
Pillow>=10.1.0
beautifulsoup4>=4.12.2
requests>=2.31.0
pyyaml>=6.0.1
pandas>=2.1.4
python-dotenv>=1.0.0
arxiv>=2.0.0
tqdm>=4.66.1
markitdown
loguru

# 数据库
sqlite3==2.6.0  # SQLite数据库支持
aiosqlite>=0.18.0  # 异步SQLite支持
diskcache
# 开发依赖
pre-commit>=3.6.2
black>=24.2.0
isort>=5.13.2
flake8>=7.0.0
flake8-docstrings>=1.7.0

# 其他依赖
streamlit>=1.42.0
loguru
pymupdf
langchain_openai
modelscope

# TODO 有关magic-pdf的转换的接口支持
# magic-pdf[full]
# --extra-index-url https://wheels.myhloli.com

# paddle 部分的依赖

# paddlepaddle==3.0.0rc0 -i https://www.paddlepaddle.org.cn/packages/stable/cpu/
# https://paddle-model-ecology.bj.bcebos.com/paddlex/whl/paddlex-3.0.0rc0-py3-none-any.whl

numpy==1.24.4